🚚 Move help section to reusable components dir
