+++ modules/ppcp-settings/resources/js/Components/Screens/Settings/Tabs/TabOverview.js
@@ -1,7 +1,7 @@
 import { __ } from '@wordpress/i18n';
 import Todos from '../Components/Overview/Todos/Todos';
 import Features from '../Components/Overview/Features/Features';
-import Help from '../Components/Overview/Help/Help';
+import Help from '../../../ReusableComponents/HelpSection';
 import { TodosHooks, CommonHooks, FeaturesHooks } from '../../../../data';
 import SpinnerOverlay from '../../../ReusableComponents/SpinnerOverlay';
 import usePaymentGatewaySync from '../../../../hooks/usePaymentGatewaySync';
